On second thought remove invariants checking
The problem with doing invariants checking in this way is that
validators who chose not to enable it will continue running, if that
makes up more than 2/3 of the voting power then validators who enable
this setting will be slashed and jailed for downtime or not submitting
signatures at no fault of their own. A better alternative would be to
integrate this directly into the binary.